requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.5.0
pandas>=2.0.0
python-dateutil>=2.8.0
click>=8.1.0
pydantic>=2.0.0
cryptography>=41.0.0
msal>=1.24.0
azure-identity>=1.15.0
msgraph-core>=0.2.2